// backend/src/common/enums/role.enum.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  DIRECTOR = 'DIRECTOR',
  GENERAL_MANAGER = 'GENERAL_MANAGER',
  HR_MANAGER = 'HR_MANAGER',
  BRANCH_MANAGER = 'BRANCH_MANAGER',
  FLOOR_MANAGER = 'FLOOR_MANAGER',
  RECEPTIONIST = 'RECEPTIONIST',
  TECHNICIAN = 'TECHNICIAN',
  STORE_MANAGER = 'STORE_MANAGER',
  ACCOUNTANT = 'ACCOUNTANT',
  MARKETING_MANAGER = 'MARKETING_MANAGER',
  DIGITAL_MARKETING = 'DIGITAL_MARKETING',
  TRAINER = 'TRAINER',
  ACADEMY_STUDENT = 'ACADEMY_STUDENT',
  FRANCHISE_OWNER = 'FRANCHISE_OWNER',
  CUSTOMER = 'CUSTOMER',
}